requirements.txt
torch==2.0.1+cu118
diffusers==0.10.2
transformers==4.27.1
pandas==1.4.3
Pillow==9.2.0
numpy==1.22.4
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.7.3
tensorflow==2.9.1
matplotlib==3.5.2
lpips==0.1.4
segmentation-models-pytorch==0.2.0
accelerate==0.11.0
datasets==2.3.2
huggingface_hub==0.8.1
requests==2.28.1
tqdm==4.64.0
